# rentalcar-api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rental Car Server",
    "description": "Api documentation rental car server",
    "version": "1.0.0",
    "contact": {
      "name": "Rafli Andreansyah",
      "email": "rafliandreansyah957@gmail.com",
      "url": "https://www.linkedin.com/in/rafli-andreansyah-86299217a"
    }
  },
  "servers": [
    {
      "url": "https://{env}:{port}/api/{version}",
      "description": "Server developer rental car",
      "variables": {
        "env": {
          "description": "Environtment url",
          "default": "localhost",
          "enum": [
            "localhost"
          ]
        },
        "port": {
          "description": "port server rental car",
          "default": "8080",
          "enum": [
            "8080",
            "8081"
          ]
        },
        "version": {
          "description": "api version",
          "default": "v1",
          "enum": [
            "v1"
          ]
        }
      }
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "get list of users",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "1"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "20"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success response get list users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataArrayObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "total_item": 50,
                      "per_page": 10,
                      "current_page": 1,
                      "last_page": 5,
                      "next_page_url": null,
                      "prev_page_url": null,
                      "status": 200,
                      "error": null,
                      "data": [
                        {
                          "id": "0092HSH-0092HSH",
                          "image_url": "https://user-profilerafli.com",
                          "email": "rafli@gmail.com",
                          "name": "rafli a",
                          "phone": "+62039477755",
                          "is_active": "true"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/AccessForbiddenError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "create data users",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "required": true,
                    "type": "string"
                  },
                  "password": {
                    "required": true,
                    "type": "string"
                  },
                  "name": {
                    "required": true,
                    "type": "string"
                  },
                  "phone": {
                    "required": true,
                    "type": "string"
                  },
                  "role": {
                    "required": true,
                    "type": "string"
                  },
                  "dob": {
                    "description": "timestamp",
                    "required": false,
                    "type": "number"
                  },
                  "image": {
                    "required": false,
                    "type": "array",
                    "items": {
                      "format": "binary",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success create user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 201,
                      "error": null,
                      "data": {
                        "id": "00KDD-7DDH8-KKS88",
                        "name": "rafli andreansyah",
                        "email": "rafli@gmail.com",
                        "image_url": null,
                        "dob": null,
                        "phone": "081232720999",
                        "role": "user",
                        "is_active": true
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/AccessForbiddenError"
          },
          "409": {
            "description": "error when user already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataObject"
                },
                "examples": {
                  "conflict": {
                    "value": {
                      "status": 409,
                      "error": "user with this email already registered",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/users/authorization": {
      "get": {
        "description": "get user car authorization",
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "1"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "20"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success get user authorization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataArrayObjectPaging"
                },
                "examples": {
                  "success": {
                    "value": {
                      "total": 50,
                      "per_page": 10,
                      "current_page": 1,
                      "last_page": 5,
                      "next_page_url": null,
                      "prev_page_url": null,
                      "status": 200,
                      "error": null,
                      "data": [
                        {
                          "id": "0092HSH-0092HSH",
                          "image_url": "https://user-profilerafli.com",
                          "name": "rafli a",
                          "email": "rafli@gmail.com",
                          "phone": "+62039477755",
                          "is_active": "true",
                          "cars_authorizations": [
                            {
                              "id": "00FYY-66DYY-666DH",
                              "name": "Toyota Fortuner",
                              "image_url": "https://fortunerfjjs888s.com",
                              "year": 2020,
                              "transmission": "Automatic",
                              "price": 250000.00
                            }
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "get detail user by id",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "response success get detail users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 200,
                      "error": null,
                      "data": {
                        "id": "8DJS-KWME-DMWA-MDMM",
                        "name": "suprianto alexandre",
                        "email": "supri@gmail.com",
                        "image_url": "https://media.istockphoto.com/id/1386479313/photo/happy-millennial-afro-american-business-woman-posing-isolated-on-white.webp?b=1&s=170667a&w=0&k=20&c=ahypUC_KTc95VOsBkzLFZiCQ0VJwewfrSV43BOrLETM=",
                        "dob": "1996-12-20",
                        "phone": "+6289373622",
                        "is_active": "true",
                        "role": "USER",
                        "date_created": "2023-10-30"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UserNotFoundError"
          },
          "403": {
            "$ref": "#/components/responses/AccessForbiddenError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "delete": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "delete user by id",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success delete user data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataString"
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": "200",
                      "error": null,
                      "data": "success delete data user"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UserNotFoundError"
          },
          "403": {
            "$ref": "#/components/responses/AccessForbiddenError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "patch": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "patch user data",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "image": {
                    "type": "array",
                    "items": {
                      "format": "binary"
                    }
                  },
                  "dob": {
                    "type": "string"
                  },
                  "is_active": {
                    "description": "only role admin",
                    "type": "string"
                  },
                  "role": {
                    "description": "only role admin",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success edit user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 200,
                      "error": null,
                      "data": {
                        "id": "00AAS-83JJD-00AAS-83JJD",
                        "name": "Jono joni",
                        "phone": "+62893874744",
                        "image": "https://imageprofile0837ssss",
                        "dob": null,
                        "is_active": true,
                        "role": "USER"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UserNotFoundError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/users/transactions": {
      "get": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "get list transaction by user id",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "1"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "20"
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "success get list transaction by user id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataArrayObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "total_item": 50,
                      "per_page": 10,
                      "current_page": 1,
                      "last_page": 5,
                      "next_page_url": null,
                      "prev_page_url": null,
                      "status": 200,
                      "error": null,
                      "data": [
                        {
                          "id": "0JJS8-884JN-MMMS8-KKSK0",
                          "no_invoice": "trx-snn89920mmdsm",
                          "start_date": "2022-10-25T10:00:00",
                          "end_date": "2022-10-26T10:00:00",
                          "duration": 1,
                          "car_name": "pajero",
                          "brand": "mitsubishi",
                          "total_price": 300000.00,
                          "status": "FINISH"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/AccessForbiddenError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/UserNotFoundError"
          }
        }
      }
    },
...
    "/transactions": {
      "get": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "tags": [
          "Transactions"
        ],
        "description": "get list of transactions an only admin can access",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "1"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "20"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success get list of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataArrayObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "total_item": 50,
                      "per_page": 10,
                      "current_page": 1,
                      "last_page": 5,
                      "next_page_url": null,
                      "prev_page_url": null,
                      "status": 200,
                      "error": null,
                      "data": [
                        {
                          "id": "0JJS8-884JN-MMMS8-KKSK0",
                          "no_invoice": "trx-snn89920mmdsm",
                          "start_date": "2022-10-25T10:00:00",
                          "end_date": "2022-10-26T10:00:00",
                          "duration": 1,
                          "car_name": "pajero",
                          "brand": "mitsubishi",
                          "total_price": 300000.00,
                          "status": "WAITING_PAYMENT"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/AccessForbiddenError"
          }
        }
      },
      "post": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "user post transactions",
        "tags": [
          "Transactions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "start_date": {
                    "type": "string",
                    "required": true
                  },
                  "duration": {
                    "type": "number",
                    "required": true
                  },
                  "user_id": {
                    "type": "string",
                    "required": true
                  },
                  "car_id": {
                    "type": "string",
                    "required": true
                  }
                }
              },
              "examples": {
                "success": {
                  "value": {
                    "start_date": "2020-08-22T10:00:00",
                    "duration": 1,
                    "user_id": "HHSY8-9JDN9-LLSJI-80KSM",
                    "car_id": "88MSS-LLLSJ-KEWM8-SKKDS"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success create transactions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 200,
                      "error": null,
                      "data": {
                        "id": "0JJS8-884JN-MMMS8-KKSK0",
                        "no_invoice": "trx-snn89920mmdsm",
                        "start_date": "2022-10-25T10:00:00",
                        "end_date": "2022-10-26T10:00:00",
                        "duration": 1,
                        "car_name": "pajero",
                        "car_image_url": "https://image.com",
                        "brand": "mitsubishi",
                        "year": 2020,
                        "capacity": 6,
                        "cc": 2500,
                        "price": 300000.00,
                        "tax": 0,
                        "discount": 0,
                        "total_price": 300000.00
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AppSecurity": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "GeneralResponseWithDataObject": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "GeneralResponseWithDataArrayObject": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "GeneralResponseWithDataArrayObjectPaging": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "per_page": {
            "type": "integer"
          },
          "current_page": {
            "type": "integer"
          },
          "last_page": {
            "type": "integer"
          },
          "next_page_url": {
            "type": "string"
          },
          "prev_page_url": {
            "type": "string"
          },
          "from": {
            "type": "number"
          },
          "to": {
            "type": "number"
          },
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "GeneralResponseWithDataString": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "UserNotFoundError": {
        "description": "Response error for user not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GeneralResponseWithDataObject"
            },
            "examples": {
              "error": {
                "value": {
                  "status": 404,
                  "error": "user doesn't exist",
                  "data": null
                }
              }
            }
          }
        }
      },
...
      "AccessForbiddenError": {
        "description": "response error forbidden access",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GeneralResponseWithDataArrayObject"
            },
            "examples": {
              "forbidden": {
                "value": {
                  "status": 403,
                  "error": "access forbidden to access this data",
                  "data": null
                }
              }
            }
          }
        }
      },
...
      "UnauthorizedError": {
        "description": "response error forbidden access",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GeneralResponseWithDataArrayObject"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "status": 401,
                  "error": "authorization required",
                  "data": null
                }
              }
            }
          }
        }
      }
    }
  }
}